루틴 상세 조회 api 호출

+++ app/src/main/java/com/makeus/milliewillie/ui/todayWorkout/TodayWorkoutFeedFragment.kt
@@ -8,13 +8,16 @@ import androidx.recyclerview.widget.RecyclerView
 import com.makeus.base.disposeOnDestroy
 import com.makeus.base.fragment.BaseDataBindingFragment
 import com.makeus.base.recycler.BaseDataBindingRecyclerViewAdapter
+import com.makeus.milliewillie.ActivityNavigator
 import com.makeus.milliewillie.R
 import com.makeus.milliewillie.databinding.FragmentTodayWorkoutFeedBinding
+import com.makeus.milliewillie.databinding.WorkoutFeedRoutineRecyclerItemBinding
 import com.makeus.milliewillie.databinding.WorkoutRoutineRecyclerItemBinding
 import com.makeus.milliewillie.model.MyRoutineInfo
 import com.makeus.milliewillie.model.TodayRoutines
 import com.makeus.milliewillie.ui.home.tab2.WorkoutFragment
 import com.makeus.milliewillie.ui.home.tab2.WorkoutFragment.Companion.EXERCISE_ID
+import com.makeus.milliewillie.ui.home.tab2.WorkoutFragment.Companion.isModifiedRoutine
 import com.makeus.milliewillie.util.Log
 import com.makeus.milliewillie.util.SharedPreference
 import io.reactivex.android.schedulers.AndroidSchedulers
@@ -22,6 +25,10 @@ import org.koin.android.viewmodel.ext.android.viewModel
 import kotlin.properties.Delegates
 
 class TodayWorkoutFeedFragment: BaseDataBindingFragment<FragmentTodayWorkoutFeedBinding>(R.layout.fragment_today_workout_feed) {
+
+    companion object {
+        const val ROUTINE_ID_KEY = "ROUTINE_ID_KEY"
+    }
 
     private val viewModel by viewModel<TodayWorkoutViewModel>()
 
@@ -72,7 +79,7 @@ class TodayWorkoutFeedFragment: BaseDataBindingFragment<FragmentTodayWorkoutFeed
             })
             adapter = BaseDataBindingRecyclerViewAdapter<MyRoutineInfo>()
                 .addViewType(
-                    BaseDataBindingRecyclerViewAdapter.MultiViewType<MyRoutineInfo, WorkoutRoutineRecyclerItemBinding>(R.layout.workout_routine_recycler_item){
+                    BaseDataBindingRecyclerViewAdapter.MultiViewType<MyRoutineInfo, WorkoutFeedRoutineRecyclerItemBinding>(R.layout.workout_feed_routine_recycler_item){
                         viFeed = this@TodayWorkoutFeedFragment
                         item = it
                     }
@@ -132,21 +139,30 @@ class TodayWorkoutFeedFragment: BaseDataBindingFragment<FragmentTodayWorkoutFeed
     fun onClickImage(status: Int) {
         when (status) {
             1 -> { // 루틴 수정 페이지
+                if (!isEdit) {
+                    ActivityNavigator.with(this).routine().apply {
+                        putExtra(ROUTINE_ID_KEY, allRoutineArray[position].routineId)
+                        isModifiedRoutine = true
+                        start()
+                    }
 
+                }
             }
             2 -> { // 루틴 삭제 API
-                viewModel.apiRepository.deleteRoutine(SharedPreference.getSettingItem(EXERCISE_ID)!!.toLong(), allRoutineArray[position].routineId)
-                    .observeOn(AndroidSchedulers.mainThread())
-                    .subscribe {
-                        if (it.isSuccess) {
-                            Log.e("deleteRoutine 호출 성공")
+                if (isEdit) {
+                    viewModel.apiRepository.deleteRoutine(SharedPreference.getSettingItem(EXERCISE_ID)!!.toLong(), allRoutineArray[position].routineId  )
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe {
+                            if (it.isSuccess) {
+                                Log.e("deleteRoutine 호출 성공")
 
-                            viewModel.removeRoutineItem(position)
-                        } else {
-                            Log.e("deleteRoutine 호출 실패")
-                            Log.e(it.message)
-                        }
-                    }.disposeOnDestroy(this)
+                                viewModel.removeRoutineItem(position)
+                            } else {
+                                Log.e("deleteRoutine 호출 실패")
+                                Log.e(it.message)
+                            }
+                        }.disposeOnDestroy(this)
+                }
             }
         }
     }
